HAI-215 Move common logic and styles for address search used in big hanke map and hanke form to own component
