+++ src/domain/map/HankeMap.tsx
@@ -1,5 +1,4 @@
 import React, { useState } from 'react';
-import { Coordinate } from 'ol/coordinate';
 import Map from '../../common/components/map/Map';
 import Controls from '../../common/components/map/controls/Controls';
 import LayerControl from '../../common/components/map/controls/LayerControl';
@@ -19,8 +18,7 @@ import HankeHoverBox from './components/HankeHover/HankeHoverBox';
 import MapGuide from './components/MapGuide/MapGuide';
 import HankkeetProvider from './HankkeetProvider';
 import MapControl from '../../common/components/map/controls/MapControl';
-import AddressSearch from './components/AddressSearch/AddressSearch';
-import CenterOnCoordinate from './components/interations/CenterOnCoordinate';
+import AddressSearchContainer from './components/AddressSearch/AddressSearchContainer';
 
 const HankeMap: React.FC = () => {
   const [zoom] = useState(9); // TODO: also take zoom into consideration
@@ -31,11 +29,6 @@ const HankeMap: React.FC = () => {
     setHankeFilterStartDate,
     setHankeFilterEndDate,
   } = useDateRangeFilter();
-  const [addressCoordinate, setAddressCoordinate] = useState<Coordinate | undefined>();
-
-  function handleAddressSelect(coordinate: Coordinate | undefined) {
-    setAddressCoordinate(coordinate);
-  }
 
   return (
     <>
@@ -46,9 +39,7 @@ const HankeMap: React.FC = () => {
       >
         <h1 className={styles.allyHeader}>Karttasivu</h1> {/* For a11y */}
         <Map zoom={zoom} mapClassName={styles.mapContainer__inner}>
-          <div className={styles.mapContainer__addressSearch}>
-            <AddressSearch onAddressSelect={handleAddressSelect} />
-          </div>
+          <AddressSearchContainer position={{ top: '1rem', left: 150 }} />
 
           <MapGuide />
           {mapTileLayers.ortokartta.visible && <Ortokartta />}
@@ -63,8 +54,6 @@ const HankeMap: React.FC = () => {
               <HankeLayer />
             </GeometryHover>
           </HankkeetProvider>
-
-          <CenterOnCoordinate coordinate={addressCoordinate} />
 
           <Controls>
             <ControlPanel className={styles.dateRangeControl}>
