+++ src/domain/map/Map.module.scss
@@ -1,5 +1,3 @@
-@import 'layout.scss';
-
 .borders {
   border: 1px solid black;
 }
@@ -13,17 +11,6 @@
     left: 0;
     bottom: 0;
     right: 0;
-  }
-
-  &__addressSearch {
-    position: absolute;
-    top: 1rem;
-    left: 150px;
-    width: 300px;
-
-    @include respond-below(s) {
-      width: auto;
-    }
   }
 }
