+++ src/domain/map/components/AddressSearch/AddressSearchContainer.module.scss
@@ -1,0 +1,11 @@
+@import 'layout.scss';
+
+.container {
+  position: absolute;
+  z-index: 1000;
+  width: 300px;
+
+  @include respond-below(s) {
+    width: auto;
+  }
+}
